Use polling instead of events for state updates
As is described in the `Watch` function, the dbus event subscription was
causing a lot of problems where it would, after a short while, start
being bombarded by events and it was *only* happening while the shim was
watching for events... as in the entire dbus seems to get bombared when
*we* listen for a short while.
Interestingly as soon as we stop our process the events stop rolling in
(visible via dbus-monitor).

This caused tremendous CPU load in the shim due to processing these
events within the dbus library.

With this polling method we never see these events being blasted on the
dbus.

Seems like there is a bug in the dbus library that is likely the culprit
of these events.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>