This rewrites most/all of the transient units
Now the shim will use real systemd units instead of transient ones. The
reason for this is because systemd purges the state of transient units
on exit
This also opens us up to using systemd for other things, such as setting
up the mount namespace, running setup/teardown jobs for the unit, and
other fun things which we'll now no longer need to react to in the
daemon and systemd will just handle for us.

It also happens to fix some bugs/edge cases that we weren't handling
well before.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>