Make exit status more robust
There are a lot of edge cases with using systemd to query state, it can
be racey among other things.

This change writes exit status to a file as an `ExecStop` to the service
unit.
Internally we exlusively rely on this data for exit status collection
rather than the systmed API.