Fix hung ExecStopPost commands due to closed stdio
When setting the stdio of the service to the client provided fifo's, we
run into a case where the client may have closed the pipes (or a pipe)
and can no longer run our neccessary cleanup commands since systemd will
block trying to open the pipes.

Now instead of using the stdio of the systemd unit, we pass those fifo's
as environment variables to the helper process we run just before
running the container which sets the pipes as the stdio of the process
just before executing runc.

This change makes it neccessary to once again run the helper command
before every container create.

Log messages from our helper will go to the journal instead of the
stderr fifo, at least for now.
I may want to change this so that we can grab such logs for debugging
without having to go through the sdjournal api.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>