Re-fix exec pid on quick exit
When an exec exits quickly (e.g. 'exit 1'), it may cause systemd to mark
the unit as failed, however if runc wrote a pid file we can go ahead and
return that because the process started.

We are not getting a proper exit code in this case (unless the real exit
code is 0), so still need to figure out a fix for that.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>