Use subreaper in startup helper to check fast exit
Since the runc command may be detaching (really it should always
detach), we'll always get an exit code of 0 (unless there was some error
related to runc), and we wind up with the wrong exit code recorded to
our fallback file.

This change introduces a process reaper so we can correctly wait for the
container to exit.

This is *only* used for processes that are short lived which wind up
messing up systemd.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>