Clean up/fix some of the reaper behavior.
Because the caller (of reap()) was already caling wait on the runc
process, handing the exit status from runc was trickier.

Some of the way the reap loop was strutured was also just incorrect
since we were waiting on signals but continuing on things like EINTR,
meanwhile we'd only ever get another signal if the contianer died.

These changes should hopefully take care of 2 of the test failures where
the tests are expeceting `NewTask()` to fail.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>