Refactor subreaper for faster response
There's no need to wait 1s (which is racey anyway because something
could exit at the 1s mark) to tell if the process is really running.

As soon as we have the pid we can turn off subreaper and check to make
sure the pid is still running.
Systemd should be able to monitor the process after.

This significantly speeds up process init since we aren't essentially
doing a 1s sleep anymore.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>