Fix warning for C write call
The idea here was to write a header with the size of our message but
really this doesn't work very well as is. The data needs to be encoded
somehow.

It looks like other error messages were not doing this anyway so just to
make things consistent for now and to silence the compiler warnings,
just write the message itself.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>